Move get_consumer_offsets out of kafka client
